+++ dspace-api/src/main/java/org/dspace/app/mediafilter/ImageMagickPdfThumbnailFilter.java
@@ -17,13 +17,30 @@ public class ImageMagickPdfThumbnailFilter extends ImageMagickThumbnailFilter {
         throws Exception
     {
 		File f = inputStreamToTempFile(source, "impdfthumb", ".pdf");
-		File f2 = getImageFile(f, 0);
-	    f.delete();
-    	File f3 = getThumbnailFile(f2);
-	    f2.delete();
-	    byte[] bytes = Files.readAllBytes(f3.toPath());
-	    f3.delete();
-	    return new ByteArrayInputStream(bytes);
+		File f2 = null;
+	    File f3 = null;
+	    try
+	    {
+		    f2 = getImageFile(f, 0);
+		    f3 = getThumbnailFile(f2);
+		    byte[] bytes = Files.readAllBytes(f3.toPath());
+		    return new ByteArrayInputStream(bytes);
+	    }
+		finally
+	    {
+		    //noinspection ResultOfMethodCallIgnored
+		    f.delete();
+		    if (f2 != null)
+		    {
+			    //noinspection ResultOfMethodCallIgnored
+			    f2.delete();
+		    }
+		    if (f3 != null)
+		    {
+			    //noinspection ResultOfMethodCallIgnored
+			    f3.delete();
+		    }
+	    }
     }
 
    public static final String[] PDF = {"Adobe PDF"};
